Use `Write-Host` to write header and footer of graph
Using this module in my PowerShell script, the header and footer never show up (although testing it independently I can get it to work).  It seems that the header and footer are falling out of the pipeline and never make it to the output.  To be consistent with the way that `Write-Graph` works, we should output the header and footer using `Write-Host` also.